Refactored some code snippets
